review feedback from buoyad
